changed functionality of the region page so that there are cards with the region images
