+++ client/src/containers/RegionsPage.js
@@ -1,36 +1,15 @@
 import { useState } from "react";
 import AnimalCard from "../components/AnimalCard";
+import RegionCard from "../components/RegionCard";
 
 const RegionsPage = ({allAnimals, allRegions}) => {
     const [regionalAnimals, setRegionalAnimals] = useState([]);
+    const [filterRegion, setFilterRegion] = useState("");
     const [listFiltered, setListFiltered] = useState(false);
     
-    const handleFilterClick = (e) => {
-        const filteredList = []
-        allAnimals.map(animal => {
-            if (animal.region === e.target.value) {
-                filteredList.push(animal)
-            }
-        })
-        setRegionalAnimals(filteredList)
-        if (!listFiltered) {
-            setListFiltered(true)
-        }
-    }
-
-    const createRegionLink = (region) => {
-        return (
-            <li className="region-title">
-                <button onClick={handleFilterClick} value={region}>
-                    {region[0].toUpperCase() + region.slice(1)}
-                </button>
-            </li>
-        )
-    }
-
     const generateRegionList = (regionList) => {
         if (allRegions.length) {
-            return regionList.map(region => createRegionLink(region))
+            return regionList.map(region => <RegionCard region={region} listFiltered={listFiltered} setListFiltered={setListFiltered} allAnimals={allAnimals} setFilterRegion={setFilterRegion} setRegionalAnimals={setRegionalAnimals}/>)
         }
     }
